Post Assembly Update
All CAD files have been updated with any changes and improvements that were made during construction/assembly.

The Documentation includes PDFs of the revised version and the original drawings and PCB files.

The PCB’s have been update with the future revisions that I would make if I was to make another arcade or update this one.

All the revision notes can be found in the Documentation folder for the individual parts.