BOM Update and Controls Update
I added all the parts to the BOM that were used when I transistioned from a Pi3 to a Pi4 Build.  Also I decided to change the IPAC config from keyboard to a gamepad, I configured Steamlink and it does not like keyboards.  The gamepad switch required no additional config in any of the other cores.