Power Button PSB and Code, Blean up files and Update to Neopixels
Added a PCB that acts as a intermediary to the power button allowing it
to be controlled via HTTP calls (I use it for Smartthings). The code for
the ESP  and device handler for smartthings is also included.  BOM
Updated also

Some files have been removed/renamed. and added a boot sequence to the
neopixel control code.